ml-engine/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
torch==2.3.0
torch-geometric==2.5.3
transformers==4.41.0
shap==0.45.0
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
pyarrow==16.0.0
networkx==3.3
redis==5.0.4
psycopg2-binary==2.9.9
pydantic==2.7.0
httpx==0.27.0
imbalanced-learn==0.12.3
lime==0.2.0.1
scipy==1.13.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
python-json-logger==2.0.7
